% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCreditUI_bottom}
\alias{DomesticCustomCreditUI_bottom}
\title{生成区域}
\usage{
DomesticCustomCreditUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
DomesticCustomCreditUI_bottom()
}
