% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCreditUI_left}
\alias{DomesticCustomCreditUI_left}
\title{请输入文件}
\usage{
DomesticCustomCreditUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
DomesticCustomCreditUI_left()
}
