% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomesticCustomCredit.R
\name{DomesticCustomCreditUI_right}
\alias{DomesticCustomCreditUI_right}
\title{生成区域}
\usage{
DomesticCustomCreditUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
DomesticCustomCreditUI_bottom()
}
