% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefundOverduePayments.R
\name{RefundOverduePaymentsUI_right}
\alias{RefundOverduePaymentsUI_right}
\title{生成区域}
\usage{
RefundOverduePaymentsUI_right()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
RefundOverduePaymentsUI_bottom()
}
